app/requirements.txt
boto==2.48.0
boto3==1.7.51
botocore==1.10.51
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
click==6.7
colour==0.1.5
cycler==0.10.0
docutils==0.14
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
gensim==3.4.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
nltk==3.3
numpy==1.14.5
pandas==0.23.1
pdfcrowd==4.3.5
pdfminer.six==20170720
pycryptodome==3.6.3
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.6.0
urllib3==1.23
visitor==0.1.3
waitress==1.1.0
Werkzeug==0.14.1
